Gitupdate - One command for all commit commands
This is a simple script that takes just two arguments - which branch to use for pulling/pushing and what is the commit message to be used. And internally it runs git pull, git add -A ., git status, git commit and git push. You can add path this to script in your ~/bash_profile file and use 'gitupdate' as any regular Linux command.